refactor: added namespaced fluent interface implementation
